Cleaned up the main test code & included stub test
Realized that I was trying to get dominant color from an older (probably deprecated) API and it wasn't working well anyway.  Simplified things to just pull the caption and any OCR text that can be read from the image.  This should be plenty to be able to make logic decisions for the student workshop exercise